requirements.txt
asgiref==3.4.1; python_version >= "3.6"
django-phonenumber-field==5.2.0; python_version >= "3.6"
django==3.2.7; python_version >= "3.6"
djangorestframework==3.12.4; python_version >= "3.5"
phonenumbers==8.12.33
pillow==8.3.2; python_version >= "3.6"
pytz==2021.1; python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.6"
